Calling Game::UpdateModel several times a frame
This is done by setting an update time, effectively changing
fElapsedTime in Game::UpdateModel to always be the same value, but
calling it the number of times required. I've done this so that
everything gets updated at smaller scales, thus making movement's
happened in smaller increments.

The requirement of smaller increments is essential due to collision
detection; it alleviates (to some degree) the possibility that one
object will move so fast as to pass through another object that it
should have collided with otherwise.