Added Xorg WM_CLASS identifiers
I did this so that my window manager (https://dwm.suckless.org/) is
able to identify it and not swallow it for the purposes of using a
debugger.

For those unaware; swallowing, in the context of a window manager, means
that when you open an executable from a terminal, and it has its own window,
the window manager hides the terminal afterwards to avoid having useless
terminals all over the place. I want to be able to see the terminal
while the executable is running for debugging.